refactor: move bbapi files from api folder to bbapi folder
- Move all bbapi* files from barretenberg/api to barretenberg/bbapi
- Update all #include paths to reference the new location
- Update files that depend on bbapi headers (api_client_ivc, cli.cpp)
