+++ barretenberg/cpp/src/barretenberg/api/api_client_ivc.cpp
@@ -1,9 +1,9 @@
 #include "api_client_ivc.hpp"
-#include "barretenberg/api/bbapi.hpp"
 #include "barretenberg/api/file_io.hpp"
 #include "barretenberg/api/get_bytecode.hpp"
 #include "barretenberg/api/log.hpp"
 #include "barretenberg/api/write_prover_output.hpp"
+#include "barretenberg/bbapi/bbapi.hpp"
 #include "barretenberg/client_ivc/client_ivc.hpp"
 #include "barretenberg/client_ivc/mock_circuit_producer.hpp"
 #include "barretenberg/client_ivc/private_execution_steps.hpp"
