+++ barretenberg/cpp/src/barretenberg/api/api_client_ivc.test.cpp
@@ -1,7 +1,7 @@
 #include "api_client_ivc.hpp"
-#include "barretenberg/api/bbapi.hpp"
-#include "barretenberg/api/bbapi_execute.hpp"
 #include "barretenberg/api/file_io.hpp"
+#include "barretenberg/bbapi/bbapi.hpp"
+#include "barretenberg/bbapi/bbapi_execute.hpp"
 #include "barretenberg/client_ivc/acir_bincode_mocks.hpp"
 #include "barretenberg/client_ivc/client_ivc.hpp"
 #include "barretenberg/client_ivc/mock_circuit_producer.hpp"
