+++ barretenberg/cpp/src/barretenberg/bb/cli.cpp
@@ -17,10 +17,10 @@
 #include "barretenberg/api/api_avm.hpp"
 #include "barretenberg/api/api_client_ivc.hpp"
 #include "barretenberg/api/api_ultra_honk.hpp"
-#include "barretenberg/api/bbapi.hpp"
 #include "barretenberg/api/gate_count.hpp"
 #include "barretenberg/api/prove_tube.hpp"
 #include "barretenberg/bb/cli11_formatter.hpp"
+#include "barretenberg/bbapi/bbapi.hpp"
 #include "barretenberg/common/thread.hpp"
 #include "barretenberg/flavor/ultra_rollup_flavor.hpp"
 #include "barretenberg/honk/types/aggregation_object_type.hpp"
