+++ barretenberg/cpp/src/barretenberg/bbapi/bbapi.hpp
@@ -6,8 +6,8 @@
  * This file includes and exports all command structures from specialized headers
  * and provides unified Command and CommandResponse types for the API.
  */
-#include "barretenberg/api/bbapi_client_ivc.hpp"
-#include "barretenberg/api/bbapi_execute.hpp"
-#include "barretenberg/api/bbapi_shared.hpp"
-#include "barretenberg/api/bbapi_ultra_honk.hpp"
+#include "barretenberg/bbapi/bbapi_client_ivc.hpp"
+#include "barretenberg/bbapi/bbapi_execute.hpp"
+#include "barretenberg/bbapi/bbapi_shared.hpp"
+#include "barretenberg/bbapi/bbapi_ultra_honk.hpp"
 #include "barretenberg/common/named_union.hpp"
