+++ barretenberg/cpp/src/barretenberg/bbapi/bbapi_execute.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "barretenberg/api/bbapi_client_ivc.hpp"
-#include "barretenberg/api/bbapi_shared.hpp"
-#include "barretenberg/api/bbapi_ultra_honk.hpp"
+#include "barretenberg/bbapi/bbapi_client_ivc.hpp"
+#include "barretenberg/bbapi/bbapi_shared.hpp"
+#include "barretenberg/bbapi/bbapi_ultra_honk.hpp"
 #include "barretenberg/common/throw_or_abort.hpp"
 #include <vector>
