+++ barretenberg/cpp/src/barretenberg/bbapi/bbapi_shared.hpp
