+++ barretenberg/cpp/src/barretenberg/bbapi/c_bind.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "barretenberg/api/bbapi_execute.hpp"
+#include "barretenberg/bbapi/bbapi_execute.hpp"
 #include "barretenberg/serialize/cbind_fwd.hpp"
 
 // Forward declaration for CBIND
